Udacity-SelfDrivingCarsNanodegree/Udacity-waymo-sensorfusion-exercises/requirements.txt
numpy
opencv-python
protobuf
easydict
pytorch
pillow
matplotlib
shapely
tqdm
open3d